+++ eslint.config.mjs
@@ -1,8 +1,10 @@
 import nextPlugin from 'eslint-config-next';
 
-export default [
+const config = [
   ...nextPlugin,
   {
     ignores: ['node_modules/**', '.next/**', 'dist/**'],
   },
 ];
+
+export default config;